+++ client/task-list/tasks/payments/index.js
@@ -25,6 +25,7 @@ import { recordEvent } from '@woocommerce/tracks';
 /**
  * Internal dependencies
  */
+import { createNoticesFromResponse } from '../../../lib/notices';
 import { getCountryCode } from '../../../dashboard/utils';
 import withSelect from '../../../wc-api/with-select';
 import { getPaymentMethods } from './methods';
@@ -125,6 +126,7 @@ class Payments extends Component {
 					} }
 					autoInstall
 					pluginSlugs={ currentMethod.plugins }
+					onResponse={ createNoticesFromResponse }
 				/>
 			),
 			isComplete: ! pluginsToInstall.length,